refactor : 지민님 리뷰 반영
