+++ src/main/java/com/ops/ops/modules/team/application/TeamCommandService.java
@@ -10,7 +10,6 @@ import static com.ops.ops.modules.file.exception.FileExceptionType.EXCEED_PREVIE
 import static com.ops.ops.modules.file.exception.FileExceptionType.NOT_WEBP_CONVERTED;
 import static com.ops.ops.modules.member.domain.MemberRoleType.ROLE_관리자;
 import static com.ops.ops.modules.member.domain.MemberRoleType.ROLE_팀장;
-import static com.ops.ops.modules.team.exception.TeamExceptionType.INVALID_AWARD_PARAMETERS;
 
 import com.ops.ops.global.util.FileStorageUtil;
 import com.ops.ops.modules.contest.application.convenience.ContestConvenience;
@@ -33,16 +32,12 @@ import com.ops.ops.modules.team.application.dto.response.TeamCreateResponse;
 import com.ops.ops.modules.team.domain.Team;
 import com.ops.ops.modules.team.domain.TeamSort;
 import com.ops.ops.modules.team.domain.dao.TeamRepository;
-import com.ops.ops.modules.team.exception.TeamException;
-
+import java.util.List;
+import java.util.Set;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -164,28 +159,6 @@ public class TeamCommandService {
         final TeamSort teamSort = teamSortConvenience.getValidateExistTeamSort();
 
         teamSort.updateSortType(request.mode());
-    }
-
-    public void updateAwardName(
-            final Long teamId, final String awardName, final String awardColor) {
-        validateAwardParameters(awardName, awardColor);
-        final Team team = teamConvenience.getValidateExistTeam(teamId);
-        team.updateAward(awardName, awardColor);
-    }
-
-    private void validateAwardParameters(final String awardName, final String awardColor) {
-        if (awardName == null && awardColor == null) {
-            return;
-        }
-
-        if (awardName != null
-                && awardColor != null
-                && !awardName.trim().isEmpty()
-                && !awardColor.trim().isEmpty()) {
-            return;
-        }
-
-        throw new TeamException(INVALID_AWARD_PARAMETERS);
     }
 
     private void checkWebpConverted(File existingFile) {
